+++ lib/constants.ts
@@ -1,7 +1,7 @@
 // SDK
 export const SDK_TITLE = 'Appwrite';
 export const SDK_TITLE_LOWER = 'appwrite';
-export const SDK_VERSION = '22.0.0';
+export const SDK_VERSION = '22.1.0';
 export const SDK_NAME = 'Command Line';
 export const SDK_PLATFORM = 'console';
 export const SDK_LANGUAGE = 'cli';
@@ -26,6 +26,10 @@ export const GITHUB_RELEASES_URL = `https://github.com/${GITHUB_REPO}/releases`;
 
 // API
 export const DEFAULT_ENDPOINT = 'https://cloud.appwrite.io/v1';
+
+// OAuth2
+export const OAUTH2_CLIENT_ID = "appwrite-cli";
+export const OAUTH2_SCOPES = "openid email profile";
 
 // Config resources
 export const CONFIG_RESOURCE_KEYS = [